conversational-ai-engine/requirements.txt
# Core Framework
fastapi==0.104.1
uvicorn[standard]==0.24.0
python-dotenv==1.0.0
pydantic==2.5.0
pydantic-settings==2.1.0

# LangChain and AI
langchain==0.1.0
langchain-anthropic==0.1.1
langchain-community==0.0.10
anthropic==0.26.0
openai==1.12.0
tiktoken==0.5.2

# Database and Caching
redis==5.0.1
motor==3.3.2
pymongo==4.6.1

# Kafka
aiokafka==0.10.0
confluent-kafka==2.3.0

# HTTP Client
httpx==0.25.2
aiohttp==3.9.1

# Production Features for Domain Schema Handler
backoff==2.2.1
circuitbreaker==1.4.0

# WebSocket
python-socketio==5.10.0
websockets==12.0

# Utilities
python-json-logger==2.0.7
prometheus-client==0.19.0
python-multipart==0.0.6

# Development
pytest==7.4.3
pytest-asyncio==0.21.1
black==23.11.0
flake8==6.1.0
mypy==1.7.1

# API Documentation
fastapi-pagination==0.12.13

# Additional for Step 2
# aioredis==2.0.1  # Removed - conflicts with redis package
starlette==0.27.0
# uuid==1.30  # Built-in module, not needed
